research_agent/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
google-genai>=0.1.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
duckduckgo-search>=5.0.0
